requirements.txt
certifi==2018.11.29
chardet==3.0.4
chromelogger==0.4.3
coingate-api==0.1.5
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-appconf==1.0.2
django-cleanup==3.0.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.0.0
django-guardian==1.4.9
django-ipware==2.1.0
django-maintenance-mode==0.13.0
django-maintenancemode==0.11.2
django-model-utils==3.1.2
django-notifications-hq==1.5.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
IPy==0.83
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
jsonpickle==1.0
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
